Move CacheUtils and CacheWarmer under the same namepsace
