+++ app/services/api/v3/import/importer.rb
@@ -1,11 +1,12 @@
 require 'db_helpers/search_path_helpers'
+require "#{Rails.root}/lib/modules/cache/warmer.rb"
+require "#{Rails.root}/lib/modules/cache/cleaner.rb"
 
 module Api
   module V3
     module Import
       class Importer
         include SearchPathHelpers
-        include CacheUtils
 
         # order matters a lot in here
         ALL_TABLES = [
@@ -84,8 +85,8 @@ module Api
               )
             end
           end
-          clear_cache
-          CacheWarmer::UrlsFile.generate
+          Cache::Cleaner.clear_all
+          Cache::Warmer::UrlsFile.generate
         rescue => e
           database_update.update_attribute(:status, Api::V3::DatabaseUpdate::FAILED)
           database_update.update_attribute(:error, e.message)
@@ -140,14 +141,6 @@ module Api
             "SELECT COUNT(*) FROM #{table}"
           )
           result.getvalue(0, 0)
-        end
-
-        def clear_cache
-          clear_cache_for_regexp_with_uri('/api/v3/', API_HOST)
-          clear_cache_for_regexp_with_uri('/content/', API_HOST)
-          Dictionary::Ind.instance.reset
-          Dictionary::Qual.instance.reset
-          Dictionary::Quant.instance.reset
         end
       end
     end
