+++ spec/workers/database_update_worker_spec.rb
@@ -16,6 +16,7 @@ RSpec.describe DatabaseUpdateWorker, type: :worker do
       allow_any_instance_of(
         Api::V3::Import::Importer
       ).to receive(:import).and_return({})
+      allow(Cache::Warmer::UrlsFile).to receive(:generate)
     end
 
     it "updates database_updates status to FINISHED" do
